345 weighted and quadratic voting values can be float (#346)
* Make the "choice" in weighted and quadratic voting as a string to double mapping
* Fix crash when displaying String value of user vote
* Chart adjustment
* Fix caching for Top voters by aVP
* Show filters if no votes are found
* Update amounts of shimmer views

+++ Goverland/Goverland/Model/Proposal.swift
@@ -315,7 +315,7 @@ struct Proposal: Decodable, Hashable, Identifiable {
                         _vote = AnyVote(base: vote)
                     }
                 case .weighted, .quadratic:
-                    if let vote = try container.decodeIfPresent(Vote<[String: Int]>.self, forKey: key) {
+                    if let vote = try container.decodeIfPresent(Vote<[String: Double]>.self, forKey: key) {
                         _vote = AnyVote(base: vote)
                     }
                 }
